chore: extract componentDidUpdate
