+++ src/components/asset-list/RecyclerAssetList/OldAssetRecyclerList.js
@@ -18,131 +18,6 @@ const StyledRecyclerListView = styled(RecyclerListView)`
   min-height: 1;
 `;
 
-const shouldComponentDidUpdate = ({
-  nextSections,
-  lastSections,
-  nextOpenFamilyTabs,
-  lastOpenFamilyTabs,
-  dataProvider,
-  scrollToOffset,
-  paddingBottom,
-  nextIsCoinListEdited,
-  lastIsCoinListEdited,
-}) => {
-  let collectibles = {};
-  let prevCollectibles = {};
-
-  nextSections.forEach(section => {
-    if (section.collectibles) {
-      collectibles = section;
-    }
-  });
-
-  lastSections.forEach(section => {
-    if (section.collectibles) {
-      prevCollectibles = section;
-    }
-  });
-
-  const bottomHorizonOfScreen =
-    ((RecyclerAssetListSharedState.rlv &&
-      RecyclerAssetListSharedState.rlv.getCurrentScrollOffset()) ||
-      0) + RecyclerAssetListSharedState.globalDeviceDimensions;
-
-  // Auto-scroll to opened family logic 👇
-  if (nextOpenFamilyTabs !== lastOpenFamilyTabs && collectibles.data) {
-    let i = 0;
-    while (i < collectibles.data.length) {
-      if (
-        nextOpenFamilyTabs[collectibles.data[i].familyName] === true &&
-        !lastOpenFamilyTabs[collectibles.data[i].familyName]
-      ) {
-        const safeIndex = i;
-        const safeCollectibles = collectibles;
-        const familyIndex = findIndex(
-          dataProvider.getAllData(),
-          function (data) {
-            return (
-              data.item?.familyName ===
-              safeCollectibles.data[safeIndex].familyName
-            );
-          }
-        );
-
-        const focusedFamilyItem = dataProvider.getAllData()[familyIndex].item;
-        const focusedFamilyHeight = ViewTypes.UNIQUE_TOKEN_ROW.calculateHeight({
-          amountOfRows: Math.ceil(Number(focusedFamilyItem.childrenAmount) / 2),
-          isFirst: false,
-          isHeader: true,
-          isOpen: true,
-        });
-
-        const startOfDesiredComponent =
-          RecyclerAssetListSharedState.rlv.getLayout(familyIndex).y -
-          AssetListHeaderHeight;
-
-        if (
-          focusedFamilyHeight <
-          RecyclerAssetListSharedState.globalDeviceDimensions
-        ) {
-          const endOfDesiredComponent =
-            startOfDesiredComponent +
-            focusedFamilyHeight +
-            AssetListHeaderHeight;
-
-          if (endOfDesiredComponent > bottomHorizonOfScreen) {
-            scrollToOffset(
-              endOfDesiredComponent -
-                RecyclerAssetListSharedState.globalDeviceDimensions,
-              true
-            );
-          }
-        } else {
-          scrollToOffset(startOfDesiredComponent, true);
-        }
-
-        break;
-      }
-      i++;
-    }
-  }
-
-  // Auto-scroll to end of the list if something was closed/disappeared 👇
-  if (
-    RecyclerAssetListSharedState.rlv &&
-    RecyclerAssetListSharedState.rlv.getContentDimension().height <
-      bottomHorizonOfScreen +
-        ViewTypes.FOOTER.calculateHeight({
-          paddingBottom: paddingBottom || 0,
-        }) &&
-    RecyclerAssetListSharedState.rlv.getCurrentScrollOffset() > 0 &&
-    (!nextIsCoinListEdited || (!lastIsCoinListEdited && nextIsCoinListEdited))
-  ) {
-    requestAnimationFrame(() => {
-      RecyclerAssetListSharedState.rlv &&
-        RecyclerAssetListSharedState.rlv.scrollToEnd({ animated: true });
-    });
-  }
-
-  // Auto-scroll to showcase family if something was added/removed 👇
-  if (
-    collectibles.data &&
-    prevCollectibles.data &&
-    collectibles.data[0]?.familyName === 'Showcase' &&
-    (collectibles.data[0]?.childrenAmount !==
-      prevCollectibles.data[0]?.childrenAmount ||
-      prevCollectibles.data[0]?.familyName !== 'Showcase')
-  ) {
-    const familyIndex = findIndex(dataProvider.getAllData(), function (data) {
-      return data.item?.familyName === 'Showcase';
-    });
-
-    const startOfDesiredComponent =
-      RecyclerAssetListSharedState.rlv.getLayout(familyIndex).y -
-      AssetListHeaderHeight;
-    scrollToOffset(startOfDesiredComponent, true);
-  }
-};
 
 export default class RecyclerAssetList extends Component {
   static propTypes = {
@@ -177,32 +52,32 @@ export default class RecyclerAssetList extends Component {
     renderAheadOffset: deviceUtils.dimensions.height,
   };
 
-  componentDidUpdate(prevProps) {
-    const {
-      isCoinListEdited: nextIsCoinListEdited,
-      sections: nextSections,
-      openFamilyTabs: nextOpenFamilyTabs,
-      dataProvider,
-      scrollToOffset,
-      paddingBottom,
-    } = this.props;
-    const {
-      isCoinListEdited: lastIsCoinListEdited,
-      sections: lastSections,
-      openFamilyTabs: lastOpenFamilyTabs,
-    } = prevProps;
-    return shouldComponentDidUpdate({
-      dataProvider,
-      lastIsCoinListEdited,
-      lastOpenFamilyTabs,
-      lastSections,
-      nextIsCoinListEdited,
-      nextOpenFamilyTabs,
-      nextSections,
-      paddingBottom,
-      scrollToOffset,
-    });
-  }
+  //componentDidUpdate(prevProps) {
+  //  const {
+  //    isCoinListEdited: nextIsCoinListEdited,
+  //    sections: nextSections,
+  //    openFamilyTabs: nextOpenFamilyTabs,
+  //    dataProvider,
+  //    scrollToOffset,
+  //    paddingBottom,
+  //  } = this.props;
+  //  const {
+  //    isCoinListEdited: lastIsCoinListEdited,
+  //    sections: lastSections,
+  //    openFamilyTabs: lastOpenFamilyTabs,
+  //  } = prevProps;
+  //  return shouldComponentDidUpdate({
+  //    dataProvider,
+  //    lastIsCoinListEdited,
+  //    lastOpenFamilyTabs,
+  //    lastSections,
+  //    nextIsCoinListEdited,
+  //    nextOpenFamilyTabs,
+  //    nextSections,
+  //    paddingBottom,
+  //    scrollToOffset,
+  //  });
+  //}
 
   // TODO: Needs to be introduced.
   //  getStableId = index => {
