+++ src/components/asset-list/RecyclerAssetList/RainbowRecyclerAssetList.tsx
@@ -19,6 +19,7 @@ import LayoutItemAnimator from './LayoutItemAnimator';
 import OldAssetRecyclerList from './OldAssetRecyclerList';
 import RecyclerAssetListSharedState from './RecyclerAssetListSharedState';
 import hasRowChanged from './hasRowChanged';
+import { usePrevious } from '@rainbow-me/hooks';
 import { deviceUtils, logger, safeAreaInsetValues } from '@rainbow-me/utils';
 
 const StyledContainer = styled(View)`
@@ -498,6 +499,144 @@ function RainbowRecyclerAssetList({
       RecyclerAssetListSharedState.rlv?.scrollToTop(false)
     );
   }, [nativeCurrency]);
+
+  const nextSections = sections;
+  const lastSections = usePrevious(sections) || sections;
+  const nextOpenFamilyTabs = openFamilyTabs;
+  const lastOpenFamilyTabs = usePrevious(openFamilyTabs) || openFamilyTabs;
+  const nextIsCoinListEdited = isCoinListEdited;
+  const lastIsCoinListEdited =
+    usePrevious(isCoinListEdited) || isCoinListEdited;
+
+  useEffect(() => {
+    let collectibles = {};
+    let prevCollectibles = {};
+
+    nextSections.forEach(section => {
+      if (section.collectibles) {
+        collectibles = section;
+      }
+    });
+
+    lastSections.forEach(section => {
+      if (section.collectibles) {
+        prevCollectibles = section;
+      }
+    });
+
+    const bottomHorizonOfScreen =
+      ((RecyclerAssetListSharedState.rlv &&
+        RecyclerAssetListSharedState.rlv.getCurrentScrollOffset()) ||
+        0) + RecyclerAssetListSharedState.globalDeviceDimensions;
+
+    // Auto-scroll to opened family logic 👇
+    if (nextOpenFamilyTabs !== lastOpenFamilyTabs && collectibles.data) {
+      let i = 0;
+      while (i < collectibles.data.length) {
+        if (
+          nextOpenFamilyTabs[collectibles.data[i].familyName] === true &&
+          !lastOpenFamilyTabs[collectibles.data[i].familyName]
+        ) {
+          const safeIndex = i;
+          const safeCollectibles = collectibles;
+          const familyIndex = findIndex(
+            dataProvider.getAllData(),
+            function (data) {
+              return (
+                data.item?.familyName ===
+                safeCollectibles.data[safeIndex].familyName
+              );
+            }
+          );
+
+          const focusedFamilyItem = dataProvider.getAllData()[familyIndex].item;
+          const focusedFamilyHeight = ViewTypes.UNIQUE_TOKEN_ROW.calculateHeight(
+            {
+              amountOfRows: Math.ceil(
+                Number(focusedFamilyItem.childrenAmount) / 2
+              ),
+              isFirst: false,
+              isHeader: true,
+              isOpen: true,
+            }
+          );
+
+          const startOfDesiredComponent =
+            RecyclerAssetListSharedState.rlv.getLayout(familyIndex).y -
+            AssetListHeaderHeight;
+
+          if (
+            focusedFamilyHeight <
+            RecyclerAssetListSharedState.globalDeviceDimensions
+          ) {
+            const endOfDesiredComponent =
+              startOfDesiredComponent +
+              focusedFamilyHeight +
+              AssetListHeaderHeight;
+
+            if (endOfDesiredComponent > bottomHorizonOfScreen) {
+              scrollToOffset(
+                endOfDesiredComponent -
+                  RecyclerAssetListSharedState.globalDeviceDimensions,
+                true
+              );
+            }
+          } else {
+            scrollToOffset(startOfDesiredComponent, true);
+          }
+
+          break;
+        }
+        i++;
+      }
+    }
+
+    // Auto-scroll to end of the list if something was closed/disappeared 👇
+    if (
+      RecyclerAssetListSharedState.rlv &&
+      RecyclerAssetListSharedState.rlv.getContentDimension().height <
+        bottomHorizonOfScreen +
+          ViewTypes.FOOTER.calculateHeight({
+            paddingBottom: paddingBottom || 0,
+          }) &&
+      RecyclerAssetListSharedState.rlv.getCurrentScrollOffset() > 0 &&
+      (!nextIsCoinListEdited || (!lastIsCoinListEdited && nextIsCoinListEdited))
+    ) {
+      requestAnimationFrame(() => {
+        RecyclerAssetListSharedState.rlv &&
+          RecyclerAssetListSharedState.rlv.scrollToEnd({ animated: true });
+      });
+    }
+
+    // Auto-scroll to showcase family if something was added/removed 👇
+    if (
+      collectibles.data &&
+      prevCollectibles.data &&
+      collectibles.data[0]?.familyName === 'Showcase' &&
+      (collectibles.data[0]?.childrenAmount !==
+        prevCollectibles.data[0]?.childrenAmount ||
+        prevCollectibles.data[0]?.familyName !== 'Showcase')
+    ) {
+      const familyIndex = findIndex(dataProvider.getAllData(), function (data) {
+        return data.item?.familyName === 'Showcase';
+      });
+
+      const startOfDesiredComponent =
+        RecyclerAssetListSharedState.rlv.getLayout(familyIndex).y -
+        AssetListHeaderHeight;
+      scrollToOffset(startOfDesiredComponent, true);
+    }
+  }, [
+    dataProvider,
+    lastIsCoinListEdited,
+    lastOpenFamilyTabs,
+    lastSections,
+    nextIsCoinListEdited,
+    nextOpenFamilyTabs,
+    nextSections,
+    paddingBottom,
+    scrollToOffset,
+  ]);
 
   return (
     <StyledContainer onLayout={onLayout}>
