+++ src/hooks/usePrevious.ts
@@ -1,7 +1,7 @@
 import { useEffect, useRef } from 'react';
 
-export default function usePrevious<T>(value: T): T {
-  const ref = useRef<T>(value);
+export default function usePrevious<T>(value: T): T | undefined {
+  const ref = useRef<T>();
 
   useEffect(() => {
     ref.current = value;
